fixup! gpg: Add OstreeGpgVerifyResult
Add a backward-compatibility disclaimer to the documentation for
ostree_gpg_verify_result_get_all().

I want to allow for extending the OstreeGpgSignatureAttr enumeration
as needed, but that would also extend the length of the GVariant tuple
returned by ostree_gpg_verify_result_get_all().

If the tuple is used properly, extending it shouldn't be an issue.
But I don't want it being regarded as an ABI break.  Document it so
we're covered.